more changes from review with shoaib
