+++ apps/transfers/enclave/Cargo.toml
@@ -37,7 +37,7 @@ tendermint-light-client = { version = "=0.38.1", default-features = false, featu
 transfers-contract = { path = "../contracts", default-features = false }
 
 # quartz
-quartz-common = { path = "../../../core/quartz-common", features=["full"]}
+quartz-common = { git = "ssh://git@github.com/informalsystems/cycles-quartz.git", features=["full"]}
 
 [dev-dependencies]
 cw-multi-test = "2.1.0"
