+++ utils/cycles-sync/Cargo.toml
@@ -9,13 +9,6 @@ repository.workspace = true
 keywords = ["blockchain", "cosmos", "tendermint", "cycles", "quartz"]
 readme = "README.md"
 
-[lib]
-path = "src/lib.rs"
-
-[[bin]]
-name = "submit"
-path = "src/bin/submit.rs"
-
 [dependencies]
 # external
 async-trait.workspace = true
@@ -41,6 +34,8 @@ cosmwasm-std.workspace = true
 
 # quartz
 cw-tee-mtcs = { path = "../../apps/mtcs/contracts/cw-tee-mtcs/" }
+wasmd-client.workspace = true
+mtcs-enclave.workspace = true
 
 [dev-dependencies]
 rand_core.workspace = true
