--- utils/cycles-sync/src/lib.rs
@@ -1,4 +1,0 @@
-pub mod cli;
-pub mod obligato_client;
-pub mod types;
-pub mod wasmd_client;
